+++ app/Http/Controllers/TicketTypeController.php
@@ -14,15 +14,24 @@ class TicketTypeController extends Controller
     //views
     function index()
     {
+        if (!auth()->user()->can("admin panel")){
+            abort(403);
+        }
         $types = ticketTypes::all();
         return view("tickettypes/ticketTypes")->with('types', $types);
     }
 
     function create(){
+        if (!auth()->user()->can("admin panel")){
+            abort(403);
+        }
         return view("tickettypes/ticketTypesCreate");
     }
 
     function edit($id){
+        if (!auth()->user()->can("admin panel")){
+            abort(403);
+        }
         $types = ticketTypes::findOrFail($id);
         return view("tickettypes/ticketTypesEdit")->with('types', $types)->with('id', $id);
     }
@@ -31,12 +40,18 @@ class TicketTypeController extends Controller
     //functions
     function Add(Request $request)
     {
+        if (!auth()->user()->can("admin panel")){
+            abort(403);
+        }
         ticketTypes::create($request->all());
         return Redirect::to('admin/ticketType');
     }
 
     function Update($id, Request $request)
     {
+        if (!auth()->user()->can("admin panel")){
+            abort(403);
+        }
         $type = ticketTypes::findOrFail($id);
         $input = $request->all();
         $type->fill($input)->save();
@@ -45,6 +60,9 @@ class TicketTypeController extends Controller
 
     function Delete($id)
     {
+        if (!auth()->user()->can("admin panel")){
+            abort(403);
+        }
         $type = ticketTypes::findOrFail($id);
         $type->delete();
         return Redirect::to('admin/ticketType');
